modify fare type form double to BigDecimal

+++ src/main/java/com/akhil/uber_backend/uber_ride/models/Ride.java
@@ -9,6 +9,7 @@ import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 import org.locationtech.jts.geom.Point;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -41,7 +42,7 @@ public class Ride {
     @Enumerated(EnumType.STRING)
     private RideStatus rideStatus;
 
-    private Double fare;
+    private BigDecimal fare;
 
     private LocalDateTime startedAt;
