+++ src/main/java/com/akhil/uber_backend/uber_ride/models/RideRequest.java
@@ -8,6 +8,7 @@ import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 import org.locationtech.jts.geom.Point;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -31,7 +32,7 @@ public class RideRequest {
     @ManyToOne(fetch = FetchType.LAZY)
     private Rider rider;
 
-    private Double fare;
+    private BigDecimal fare;
 
     @Enumerated(EnumType.STRING)
     private PaymentMethod paymentMethod;
